refactor: change methods to GetAllCommunitiesWithMembers and GetCommunitiesByCategoryIdAndCriteria to feature Community
