+++ Coinnova.API/Controllers/CategoryController.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.Mvc;
 namespace Coinnova.API.Controllers;
 
 [ApiController]
-[Authorize]
+[Authorize(Roles = "standard")]
 [Route("api/[controller]")]
 public class CategoryController : ControllerBase
 {
@@ -21,12 +21,5 @@ public class CategoryController : ControllerBase
     {
         var categories = await _categoryService.GetCategories();
         return Ok(categories);
-    }
-
-    [HttpGet("{id}")]
-    public async Task<IActionResult> GetCommunitiesByCategoryIdAndCriteria([FromRoute] int id, [FromQuery] string criteria, [FromQuery] int skip, [FromQuery] int take)
-    {
-        var response = await _categoryService.GetCommunitiesByCategoryIdAndCriteria(id, criteria, skip, take);
-        return Ok(response);
     }
 }
