+++ Coinnova.Application/Interfaces/ICommunityService.cs
@@ -1,3 +1,5 @@
+using Coinnova.Application.Dtos.Category;
+using Coinnova.Application.Dtos.Common;
 using Coinnova.Application.Dtos.Community;
 
 namespace Coinnova.Application.Interfaces;
@@ -5,4 +7,8 @@ namespace Coinnova.Application.Interfaces;
 public interface ICommunityService
 {
     Task<List<CommunityGetDto>> Get5PopularCommunities();
+    Task<PagedResponseDto<CommunityWithMembersDto>> GetCommunitiesByCategoryIdAndCriteria(int id, string criteria,
+        int skip, int take);
+
+    Task<PagedResponseDto<CommunityWithMembersDto>> GetAllCommunitiesWithMembers(int skip, int take);
 }
