+++ Coinnova.Application/Services/CategoryService.cs
@@ -21,35 +21,4 @@ public class CategoryService : ICategoryService
         var response = categories.Adapt<IEnumerable<CategoryResponseDto>>();
         return response;
     }
-
-    public async Task<PagedResponseDto<CommunityWithMembersDto>> GetCommunitiesByCategoryIdAndCriteria(int id, string criteria, int skip, int take)
-    {
-        var query = await _unitOfWork.Categories.GetQueryCommunitiesByCategoryId(id);
-        var communities = query.Adapt<List<CommunityWithMembersDto>>();
-
-        if (criteria == "popular")
-        {
-            communities = communities.OrderByDescending(c => c.Members).ToList();
-        }
-
-        else if (criteria == "new")
-        {
-            communities = communities.OrderByDescending(c => c.CreatedAt).ToList();
-        }
-        
-        var totalCommunities = communities.Count;
-        
-        var paginated = communities.Skip(skip).Take(take).ToList();
-        
-        var hasMore = totalCommunities > (skip + take);
-        
-        return new PagedResponseDto<CommunityWithMembersDto>
-        {
-            Items = paginated,
-            HasMore = hasMore,
-            TotalCount = totalCommunities
-        };
-        
-    }
-
 }
