+++ Coinnova.Infrastructure/Repositories/CommunityRepository.cs
@@ -29,4 +29,53 @@ public class CommunityRepository : Repository<Community>, ICommunityRepository
 
         return communities;
     }
+
+    public Task<IQueryable<object>> GetQueryCommunitiesByCategoryId(int id)
+    {
+        var query =  _context.CommunityCategory
+            .Where(cc => cc.IdCategory == id)
+            .Select(cc => cc.IdCommunity)
+            .Distinct()
+            .Join(
+                _context.Community,
+                communityId => communityId,
+                c => c.Id,
+                (communityId, c) => new { Community = c }
+            )
+            .GroupJoin(
+                _context.CommunityMember,
+                c => c.Community.Id,
+                cm => cm.IdCommunity,
+                (c, members) => new 
+                {
+                    Id = c.Community.Id,
+                    Name = c.Community.Name,
+                    Description = c.Community.Description,
+                    ImageUrl = c.Community.Imageurl,
+                    CreatedAt = c.Community.Createdat,
+                    Members = members.Count()
+                });
+
+        return Task.FromResult<IQueryable<object>>(query);
+    }
+
+    public Task<IQueryable<object>> GetQueryCommunitiesWithMembers()
+    {
+        var query = _context.Community
+            .GroupJoin(
+                _context.CommunityMember,
+                c => c.Id,
+                cm => cm.IdCommunity,
+                (c, members) => new 
+                {
+                    Id = c.Id,
+                    Name = c.Name,
+                    Description = c.Description,
+                    ImageUrl = c.Imageurl,
+                    CreatedAt = c.Createdat,
+                    Members = members.Count()
+                });
+
+        return Task.FromResult<IQueryable<object>>(query);
+    }
 }
